17.asaguiar

+++ database/migrations/2025_08_26_091801_add_role_to_users_table.php
@@ -21,7 +21,9 @@ class AddRoleToUsersTable extends Migration
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');
+            if (Schema::hasColumn('users', 'role')) {
+                $table->dropColumn('role');
+            }
         });
     }
 }
